requirements.txt
requests
pillow
tkcalendar
python-decouple
